+++ src/cint.js
@@ -21,13 +21,13 @@ cint = (function() {
 		)
 	}
 
-	/** Returns true if the string starts with the given substring. 
+	/** Returns true if the string starts with the given substring. Returns false if the substring is the empty string.
 		@memberOf module:cint#
 		@param {String} str
 		@param {String} sub The substring.
 	*/
 	function startsWith(str, sub){
-		return (str.indexOf(sub) === 0)
+		return str.indexOf(sub) === 0 && sub !== ''
 	}
 
 	/** Returns the substring before the first instance of the given delimiter.